# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Council Asset & Compliance Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /my-tenant:
    get:
      operationId: getMyTenant
      tags: [users]
      summary: Get the current user's tenant workspace details
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateMyTenant
      tags: [users]
      summary: Update the current tenant's workspace settings (tenant_admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantSettingsInput"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (Super Admin)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantList"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant (Super Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant details
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{tenantId}/suspend:
    post:
      operationId: suspendTenant
      tags: [tenants]
      summary: Suspend a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant suspended
  /tenants/{tenantId}/activate:
    post:
      operationId: activateTenant
      tags: [tenants]
      summary: Activate a suspended tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant activated
  /tenants/{tenantId}/stats:
    get:
      operationId: getTenantStats
      tags: [tenants]
      summary: Get per-tenant usage statistics
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantStats"

  # ── USERS ─────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in current tenant
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: role
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user in current tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user detail
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/deactivate:
    post:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User deactivated

  # ── ASSETS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TenantSettingsInput:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        notificationsEnabled:
          type: boolean
        notificationEmail:
          type: ["string", "null"]
    Tenant:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, trial]
        contactEmail:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        maxAssets:
          type: ["integer", "null"]
        maxUsers:
          type: ["integer", "null"]
        notificationsEnabled:
          type: boolean
        notificationEmail:
          type: ["string", "null"]
        assetCount:
          type: ["integer", "null"]
        userCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        plan:
          type: string
        maxAssets:
          type: integer
        maxUsers:
          type: integer
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        plan:
          type: string
        maxAssets:
          type: integer
        maxUsers:
          type: integer
        status:
          type: string
    TenantList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tenant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TenantStats:
      type: object
      required: [tenantId, assetCount, userCount]
      properties:
        tenantId:
          type: string
        assetCount:
          type: integer
        userCount:
          type: integer
        complianceItemCount:
          type: integer
        overdueCount:
          type: integer
        lastImportAt:
          type: ["string", "null"]

    # Users
    User:
      type: object
      required: [id, username, email, role, isActive, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, tenant_admin, compliance_manager, team_member, auditor]
        isActive:
          type: boolean
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [username, email, password, role]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [tenant_admin, compliance_manager, team_member, auditor]
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
    UserList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Assets
...